New interactive map and graph of imagery availability by time
I have created a new interactive map on Leaflet to show in a
more fancy way which areas DigitalGlobe covers over the Torres
Strait. Also, I have included a graph which groups the timestamps
by month/day and show which dates are the most frequent.

TODO:
- Change colors/borders polygons
- Create a good set-up to start playing with AIS data and Sat data
- Create a new function to do all this behind courtains - including dowload

BUGS:

- If you want to run this on local, remember to create the .gbdx credential
profile. GBDX documentation can be followed to create this.