Intersection between GBDX data and AIS data
Two new IPythonNotebooks to visualize and intersect GBDX data and
AIS data. The first notebook, gbdx_data_description makes a conec-
-tion with the GBDX API and allow to download the imagery metadata
and assess the coverage of these images over the selected zone. A
.csv is generated with all the data.

The second notebook, gbdx_ais_intersection, reads the AIS data from
a POSTGRESQL data base and use the POSTGIS functions to find an
intersection between the imagery footprint and the vessels locations
A second step is to retrieve the imagery corresponding to this over-
-lap.

One last thing was to change names of files to make them more
readable.

TODO

- Finish pipeline to download images to instance
- Implement rasterio to open images and label data with AIS

BUGS
-  You have to create a database.py file with a dict to connect to
   the DB in POSTGRES using sqlalchemy.
-  To interact with the GBDX data we have to create a .gbdx-auth
   file. The gbdxtools documentation explain how to do this.
-  For now, the GBDX API is not correctly downloading the data
   into the user S3 bucket. That's a shame! :grin: