Intersection between GBDX and AIS - Label dataset
This new commit includes new changes into the gbdx_ais_intersection
notebook With this changes we can retrieve cropped sat imagery which
contains only vessels. This will allow us to create a training data
with a set of labeled vessels that can serve as a first step into a
model to identify other vessels, even with their AIS off. A key feature
of this process is the use of GBDX servers instead of the AWS instance.

TODO

- Get panchromatic images
- Cover more area on the satellite imagery
- Create a better apprach to buffer creation - at least calibrate

BUGS

- To run this code you will need gbdx access using the gbdxtools and
gbdx_auth libraries. More info can be consulted on their webpage.

- Only BN images. Bug in code (already mail them)