Feature generation from AIS: Distance from shore  :triangular_ruler:
This coomit includes changes into the gbdx_ais_intersection notebook
, and a new notebook that create a function to calculate the distance
to the neirest point of a shore. The function takes a dataframe as an
input, and requires to specify the names of the longitude and latitude
variables. The output is a new dataframe equal to the input one, but
with a new variable: distance_to_shore in kilometers.

To calculate this distances, rather than read a .shp of shores, the
function uses matplotlib Basemap library to extract the coastlines
poinst and use that as an input of the distance calculation. This
process can take time (~30 segs). For that reason, the code saves this
coodinates in a new folder call "data" inside your PATH in a .npy
format.

TODO:
- Find a way of create a .temp file instead of a .npy local file
- Create an option to take an np.array as input and as output.
  This depends heavily in the pipeline definition

BUGS:

- As all notebooks (code) related to gbdx data, you need an
  authentication file.