Add a script to retrieve satellite imagery with Planet :satellite:
This .py script run a request of metadata using the Planet API.
This is only a test to retrieve images for the city of Rotterdam
(Netherlands). The use of this code is fairly easy, relies basically
on a GeoJson file that creates an AOI to retrieve images. Unlike GBDX,
this Planet API works in a more rustic way. You'll have to specify
every query in a separate way and it will retrieve a .json dict with
all the metadata.