Changes to GBDX data exploration :grimacing:
This new changes create queries to retrieve data from all the marine
areas in the world. This will be important since allow us to have a
better coverage of the areas where fishing vessels are (i.e. near the
shores and inside EEZ's). The gbdx_data_description contains a function
that will retrieve all imagery looping in each marine area and will
try-catch exceptions when the limit of 1000 is exceeded.

The second notebook: 2.0-im-ais_gbdx_intersection run the previous codes
to upload data to the db instance and run instersection and geom queries
from postgresql.

TO-DO: Improve SQL queries, they are taking >1day
BUGS: Nothing in SQL is running as it should.