GBDX retrieve functions to marine areas and ocean areas
This new commit includes new functions to retrieve
metadata of imagery from GBDX on the whole ocean.
The function will retrieve geojson areas from CartoDB
data repository and will use them to iterate in time and
space and retrieve all the available images. The output is
a .txt file with a series of geojson strings that can be
proccessed again into Python and converted to a .csv that
can be uploaded to Postgresql and create the overlapping of
AIS and GBDX.

On this side, the notebook has been updated and soon it will
be deprecated to open space to new functions that will run as
modules :wink: