Scripts to retrieve more GBDX data: Marine areas and Oceans
This new commit includes .py files that allow to retrieve
new image metadata from the whole ocean. This code will
make request to GBDX API several times trying to break the
1000 image limit. The code takes CARTODB .geojson databases
of marine areas and oceans to use them as AOI's. This code
will create a .txt string of json dicts that can be readed
again in Python and exported again as .csv's. A Postgresql
solution is still wanting, but with the use of SQLAlchemy
we can upload directly from Python and Pandas.

TODO: Postgresql integration
BUGS: Remember that the .gbdx_auth must be updated in the root
folder (~) of your machines