First  pipeline for is_fishing model scripts for complete feature generation, training, validation, and testing of is_fishing model. Training data is preprocessed by downloading from the db and performing operations in pandas such as to drop nulls, and remove duplicate columns, additional features can then be created from the feature_generation module, such as sun_height and day_or_night. The data is then copied to the db for persistence.
ML modelling is configured in the ml_config file where input
parameters are given as a dict. These will form a row in the
db for persistence. The ml model can then be trained and a
precision score output, further metrics will be useful. The model
is pickled to the /models folder.

The predict_is_fishing model unpickles the ml model and performs
predictions on chunks from the db. The output is then uploaded
to a predictions table.

This full pipeline has been tested to confirm that it works
although since the distance_from_shore and distance_from_shore
functions are not working correctly all outputs should be ignored
until this is resolved.

BUGS:
   - EEZ defined in km not nautical miles
   - distance_from_shore and distance_to_shore multiply defined

TO DO:
    - document code and comment
    - configure evaluation db table and model params table
    - parallelize chunking for prediction
    - include more models and input params for ml_config
    - include more evaluation metrics in training