Add fantabulous shiny_app! :sparkles:
This new branch will be a development branch to a R Shiny app that will
be interact with a component list on Postgresql and create interactive
list/map to play with. This development is out of the Python pipeline
and in order to be run, you need to install all the framework, which
includes R, RStudio, RShiny and other dependencies in the code.

TODO: Integrate this app from the local setting to the AWS remote one
BUGS: None :wink: