Planet imagery retrival and process (crop and plot)
Under notebooks/planet there are new scripts created to get images from
planet. This notebook will be changed in order to integrate new
functions for planet data. As with GBDX, the process is the same: we
open a SQL conection to get the intersection between AIS and the
satelite images, and then retrieve them using the Planet API.

TODO:
- Take functions from planet and integrate them in our module
- Create a visualization pipeline to both GBDX and Planet

BUGS:
- Planet API can be buggy. To options are coded in the
notebook/planet/planet_check notebook. The first is using their clip and
ship module, and the second only download the image and crop it using
Mapbox's rasterio.