Web-app: Tool to save our oceans :dolphin:
This commit updates the code to run the shiny-app. This app visualizes
the data from the modelling, and show different components to assess
risky behaviors of individual vessels. The shown data is anonymized
and a replacement for the ID was created. The data is filtered by satellite
coverage and number of reported positions and the ranked list shows
each components sorted by percentile in the empirical distribution.

The app is currently running in a separate instance.