Data processing with Planet images
This notebook include the necessary code to retrieve images with
intersection with AIS. The notebook read from Postgresql the data and
uses the catalogID's of the images to retrieve them, download them to
the remote instance and crop them using rasterio. This notebook runs
using the Python Planet API (better that the request REST API).

This images are only for the Torres Strait.

TODO:
- Integrate the functions to src/sat_imagery
- Try to convert the geoms instead of re-project the images (take too
much time)