Update planet search metadata
This commit removes the hardcoded area of interest to use with the
Planet API and instead reads the file coords.json which is a dictionary
with coordinates in geojson format. This allows to search in any AOI by
only changing the coords.json file and yield a .csv with the imagery
metadata.